% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{does_player_cooperate}
\alias{does_player_cooperate}
\title{Does \code{player} cooperate with the other player on this round of the game?}
\usage{
does_player_cooperate(player, previous_round)
}
\arguments{
\item{player}{tbl row of the player whose cooperation we are checking}

\item{previous_round}{tbl of the previous round of the game}
}
\value{
logical of T = cooperate, F = defect
}
\description{
Does \code{player} cooperate with the other player on this round of the game?
}
\details{
Uses a version of forgiving tit-for-tat. Basic strategy is to copy
the partner's previous move, but with a chance to change. Where the partner
cooperated, players defect based on their sneakiness. Where the partner
defected, players cooperate based on their forgivingness.
}
