% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{.onLoad}
\alias{.onLoad}
\title{Set options for working with Silly Smile Sim}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{a character string giving the library directory where the
package defining the namespace was found.}

\item{pkgname}{a character string giving the name of the package.}
}
\description{
Set options for working with Silly Smile Sim
}
