% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faces.R
\name{event_facial_response}
\alias{event_facial_response}
\title{Return a target facial expression for an event}
\usage{
event_facial_response(event)
}
\arguments{
\item{event}{tbl row of the event to which we are responding}
}
\value{
tbl of feature-value pairs describing a facial expression
}
\description{
Return a target facial expression for an event
}
