% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{feature_plot}
\alias{feature_plot}
\title{Print a neat EEG-style plot of how facial features are moving throughout the
trials}
\usage{
feature_plot(feature_data, features = FEATURES)
}
\arguments{
\item{feature_data}{facial feature data to plot}

\item{features}{facial features to include in the plot}
}
\value{
ggPlot object with a feature_data plot
}
\description{
Print a neat EEG-style plot of how facial features are moving throughout the
trials
}
