% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{frames_to_ms}
\alias{frames_to_ms}
\title{Find the millisecond at which frames begin}
\usage{
frames_to_ms(frameNumbers, fps = 30)
}
\arguments{
\item{frameNumbers}{frames to get the beginnings of}
}
\value{
millisecond at which each frame in frames begins
}
\description{
Find the millisecond at which frames begin
}
