% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faces.R
\name{generate_resting_face}
\alias{generate_resting_face}
\title{Generate a random resting face}
\usage{
generate_resting_face(seed, means, sds)
}
\arguments{
\item{seed}{used to generate the random numbers}

\item{means}{single number or vector of means for each feature}

\item{sds}{single number or vector of standard deviations for each feature}
}
\value{
tbl of feature-value pairs describing the log-odds that resting face matches a feature expression
}
\description{
Generate a random resting face
}
