% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{get_decision_time}
\alias{get_decision_time}
\title{Convert a standardized decision time to a real one in ms}
\usage{
get_decision_time(z_times, M, SD, cap = TRUE)
}
\arguments{
\item{z_times}{vector of standardized times}

\item{M}{mean decision time}

\item{SD}{standard deviation of decision time}

\item{cap}{whether to cap the resulting decision times to minimum 1sd ms}
}
\value{
vector of decision times in ms
}
\description{
Convert a standardized decision time to a real one in ms
}
