% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{get_outcome_description}
\alias{get_outcome_description}
\title{Return a string describing the outcome}
\usage{
get_outcome_description(a_cooperates, b_cooperates)
}
\arguments{
\item{player_a_decision}{whether player a cooperates}

\item{player_b_decision}{whether player b cooperates}
}
\value{
string describing the outcome
}
\description{
Return a string describing the outcome
}
