% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_player_by_id}
\alias{get_player_by_id}
\title{Return the player whose id matches id}
\usage{
get_player_by_id(id, players)
}
\arguments{
\item{id}{id of the player to fetch}

\item{players}{list of players}
}
\value{
player or NA
}
\description{
Return the player whose id matches id
}
