% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ms_to_frames}
\alias{ms_to_frames}
\title{Convert a duration in milliseconds to a frame count}
\usage{
ms_to_frames(ms, fps = 30, floor = T)
}
\arguments{
\item{ms}{milliseconds}

\item{fps}{frame rate}

\item{floor}{whether to return the answer as maximum number of frames in the duration (T) or exact answer including partial frames (F)}
}
\value{
frames in \code{ms}
}
\description{
Convert a duration in milliseconds to a frame count
}
