% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\docType{data}
\name{OUTCOME_NAMES}
\alias{OUTCOME_NAMES}
\title{Names of the outcomes listed in order so that matrix(X, 2, 2) gives the
outcome matrix with 1 = steal, 2 = split}
\format{
An object of class \code{character} of length 4.
}
\usage{
OUTCOME_NAMES
}
\description{
Names of the outcomes listed in order so that matrix(X, 2, 2) gives the
outcome matrix with 1 = steal, 2 = split
}
\keyword{datasets}
