% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{simulate_behavioural_markers}
\alias{simulate_behavioural_markers}
\title{Simulate behavioural data}
\usage{
simulate_behavioural_markers(
  player,
  partner,
  n_rounds,
  decision_time_mean = 750,
  decision_time_sd = 100
)
}
\arguments{
\item{player}{tbl row of the player}

\item{partner}{tbl row of the player's partner (another player)}

\item{n_rounds}{number of rounds to play}

\item{decision_time_mean}{mean decision time of the sample}

\item{decision_time_sd}{standard deviation of the above}
}
\value{
tbl with $id, $round_id, and markers for $decision_time, $partner_decision_time, $reveal_time, and $round_end_time all in ms.
}
\description{
Simulate behavioural data
}
