% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faces.R
\name{simulate_faces}
\alias{simulate_faces}
\title{Simulate facial data for players}
\usage{
simulate_faces(players, behavioural_data, ms_between_expressions = 150)
}
\arguments{
\item{players}{list of players in the game}

\item{behavioural_data}{tbl of behavioural data (player, round, and temporal event markers)}

\item{ms_between_expressions}{milliseconds between expressions}
}
\value{
tbl of a simulated round with a column for each feature and a row for each frame
}
\description{
Simulate facial data for players
}
