% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{simulate_game}
\alias{simulate_game}
\title{Simulate the cooperation/defecting behaviour in the game}
\usage{
simulate_game(player_a, player_b, n_rounds)
}
\arguments{
\item{player_a}{player taking the role of player a}

\item{player_b}{player taking the role of player b}

\item{n_rounds}{number of rounds to simulate}
}
\value{
tbl with $round_id, $player_a_id, $player_b_id, $player_a_cooperates, $player_b_cooperates, $outcome
}
\description{
Simulate the cooperation/defecting behaviour in the game
}
