% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players.R
\name{simulate_players}
\alias{simulate_players}
\title{Simulate the underlying tendencies for players}
\usage{
simulate_players(n_players)
}
\arguments{
\item{n_players}{number of players to simulate}
}
\value{
tbl of $id, $z_decision_speed, $forgivingness, $sneakiness
}
\description{
Simulate the underlying tendencies for players
}
