% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{simulate_rounds}
\alias{simulate_rounds}
\title{Simulate the data}
\usage{
simulate_rounds(players, n_rounds, fps = 30)
}
\arguments{
\item{players}{tbl of players}

\item{n_rounds}{number of rounds to simulate}

\item{fps}{frame rate of simulations}
}
\value{
tbl of experiment data
}
\description{
Simulate the data
}
