% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faces.R
\name{update_face}
\alias{update_face}
\title{Update a face by moving some amount from the current value to the target}
\usage{
update_face(face)
}
\arguments{
\item{face}{face to update}
}
\value{
updated \code{face}
}
\description{
Update a face by moving some amount from the current value to the target
}
