% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faces.R
\name{simulate_faces}
\alias{simulate_faces}
\title{Simulate facial data for players}
\usage{
simulate_faces(behavioural_data, ms_between_expression = 150)
}
\arguments{
\item{behavioural_data}{tbl of behavioural data (player, round, and temporal event markers)}
}
\value{
tbl of a simulated round with a column for each feature and a row for each frame
}
\description{
Simulate facial data for players
}
